added a new componet that shows a group of datasets
