+++ app/assets/javascripts/dashboard/gallery_dataset_view.js
@@ -2,20 +2,15 @@
 /* eslint-disable jsx-a11y/href-no-hash */
 import * as React from "react";
 import { connect } from "react-redux";
-import { Row, Col, Modal, Card, Dropdown, Menu, Icon } from "antd";
+import { Modal } from "antd";
 import Utils from "libs/utils";
-import Markdown from "react-remarkable";
-import TemplateHelpers from "libs/template_helpers";
 import messages from "messages";
 import { createExplorational } from "admin/admin_rest_api";
+import DatasetPanel from "dashboard/dataset_panel";
 
 import type { DatasetType } from "dashboard/dataset_view";
 import type { OxalisState } from "oxalis/store";
 import type { APIUserType } from "admin/api_flow_types";
-
-const padding = 16;
-const thumbnailDimension = "500";
-const columnSpan = { xs: 24, sm: 24, md: 24, lg: 12, xl: 12, xxl: 8 };
 
 type StateProps = {
   activeUser: ?APIUserType,
@@ -25,8 +20,6 @@ type Props = {
   datasets: Array<DatasetType>,
   searchQuery: string,
 } & StateProps;
-
-// owningOrganization
 
 class GalleryDatasetView extends React.PureComponent<Props> {
   createTracing = async (
@@ -70,111 +63,26 @@ class GalleryDatasetView extends React.PureComponent<Props> {
     return sortedDatasets;
   }
 
-  renderCard(dataset: DatasetType) {
-    let description;
-    if (dataset.description) {
-      description = (
-        <Markdown
-          source={dataset.description}
-          options={{ html: false, breaks: true, linkify: true }}
-        />
-      );
-    } else {
-      description = dataset.hasSegmentation ? (
-        <p>Original data and segmentation</p>
-      ) : (
-        <p>Original data</p>
-      );
-    }
-
-    const menu = (
-      <Menu>
-        <Menu.Item key="existing">
-          <a
-            href="#"
-            onClick={() => this.createTracing(dataset, "volume", true)}
-            title="Create Volume Tracing"
-          >
-            Use Existing Segmentation Layer
-          </a>
-        </Menu.Item>
-        <Menu.Item key="new">
-          <a
-            href="#"
-            onClick={() => this.createTracing(dataset, "volume", false)}
-            title="Create Volume Tracing"
-          >
-            Use a New Segmentation Layer
-          </a>
-        </Menu.Item>
-      </Menu>
-    );
-
-    const volumeTracingMenu = (
-      <Dropdown overlay={menu} trigger={["click"]}>
-        <a href="#" title="Create Volume Tracing">
-          <img src="/assets/images/volume.svg" alt="Volume" />
-        </a>
-      </Dropdown>
-    );
-
-    return (
-      <Card bodyStyle={{ padding: 0 }} className="spotlight-item-card">
-        <span
-          className="dataset-thumbnail"
-          style={{
-            background: `url(${
-              dataset.thumbnailURL
-            }?w=${thumbnailDimension}&h=${thumbnailDimension})`,
-            backgroundSize: "cover",
-          }}
-        >
-          <div className="dataset-thumbnail-buttons">
-            <a href={`/datasets/${dataset.name}/view`} title="View Dataset">
-              <Icon type="eye-o" className="view-button" />
-            </a>
-            <a
-              href="#"
-              title="Create skeleton tracing"
-              onClick={() => this.createTracing(dataset, "skeleton", false)}
-            >
-              <img src="/assets/images/skeleton.svg" alt="Skeleton" />
-            </a>
-            {volumeTracingMenu}
-          </div>
-        </span>
-        <div className="dataset-description">
-          <div className="description-flex">
-            <h3>
-              {dataset.displayName != null && dataset.displayName !== ""
-                ? dataset.displayName
-                : dataset.name}
-            </h3>
-            <div className="dataset-description-body">
-              <p>Scale: {TemplateHelpers.formatScale(dataset.dataSource.scale)}</p>
-              {description}
-            </div>
-          </div>
-        </div>
-      </Card>
-    );
-  }
-
   render() {
+    const datasets = Utils.filterWithSearchQueryAND(
+      this.props.datasets.filter(ds => ds.isActive),
+      ["name", "description"],
+      this.props.searchQuery,
+    ).sort(Utils.localeCompareBy("formattedCreated", false));
+    if (this.props.datasets.length > 0) {
+      console.log("yay");
+    }
+    const sortedDatasets = this.sortByOrganisation(datasets);
     return (
-      <Row gutter={padding}>
-        {Utils.filterWithSearchQueryAND(
-          this.props.datasets.filter(ds => ds.isActive),
-          ["name", "description"],
-          this.props.searchQuery,
-        )
-          .sort(Utils.localeCompareBy("formattedCreated", false))
-          .map(ds => (
-            <Col className="gallery-dataset-col" {...columnSpan} key={ds.name}>
-              {this.renderCard(ds)}
-            </Col>
-          ))}
-      </Row>
+      <React.Component>
+        {sortedDatasets.map(ds => (
+          <DatasetPanel
+            key={ds.owningOrganization}
+            owningOrganization={ds.owningOrganization}
+            datasets={ds.datasets}
+          />
+        ))}
+      </React.Component>
     );
   }
 }
